// src/types/article/StructuredBlockType.ts
export enum StructuredBlockType {
  Component = 'component-block',
  RichText = 'rich-text-block',
  Image = 'image-block',
  VimeoEmbed = 'vimeo-embed-block',
  YoutubeEmbed = 'youtube-embed-block',
  Divider = 'divider-block',
  Date = 'date-block',
  Header = 'header-block',
  HeaderImage = 'header-image-block',
  Video = 'video-block',
  Quote = 'quote-block',
  Code = 'code-block'
}